Business Choices
These files outline the functionality of the inventory management application. The Application class (Application.java) contains the main structure of the application whereas the Logic and Menu classes (Logic.java/Menu.java) contain methods used in the application such as, creating a customer in the database or calculating the cost of an order.