Database table models
The three classes in this models file (Boo.java, Customer.java, Order.java) contain methods that enable the Create and Read SQL statements to be executed directly in SQL when called. The Order.java file includes an addOrder() method which is different from the createOrder() method. The addOrder() method will be used to calculate a cost for an order.  The createOrder() will be scrapped when I next edit this file.