Database design choices
This file includes justifications for the customer, book, order and order_books tables in my database. I updated the first three tables with auto increment for the ID fields. I also added the field author_name to the books table because I scaled back the database design as I had made an author_book and authors table to account for the many to many relationship between authors and books.